requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.2.0
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.51.0
propcache==0.3.0
pyee==12.1.1
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xlrd==2.0.1
yarl==1.18.3
